Add EAP6.2, 6.3 & 6.4 transformers for transaction subsystem
