+++ transactions/src/test/resources/org/jboss/as/txn/subsystem/full-1.2.xml
