+++ transactions/src/test/resources/org/jboss/as/txn/subsystem/full-3.0.xml
