+++ transactions/src/test/resources/org/jboss/as/txn/subsystem/jdbc-store.xml
