+++ transactions/src/test/resources/org/jboss/as/txn/subsystem/minimal.xml
