+++ transactions/src/test/resources/org/jboss/as/txn/subsystem/subsystem.xml
